Rename BandedGraph to LatticeGraph and delete old LatticeGraph which was a simpler cases of BandedGraph where n_blocks=1.
